memtable

+++ pkg/utils/hash.go
@@ -8,7 +8,7 @@ func Hash(data []byte, seed uint32) uint32 {
 	var (
 		m uint32 = 0xc6a4a793
 		r uint32 = 24
-		h uint32 = seed ^ (len(data) * m)
+		h uint32 = seed ^ (uint32(len(data)) * m)
 	)
 
 	i := 4
@@ -21,11 +21,11 @@ func Hash(data []byte, seed uint32) uint32 {
 
 	switch len(data) - i {
 	case 3:
-		h += uint8(data[len(data)-1]) << 16
+		h += uint32(data[len(data)-1]) << 16
 	case 2:
-		h += uint8(data[len(data)-1]) << 8
+		h += uint32(data[len(data)-1]) << 8
 	case 1:
-		h += uint8(data[len(data)-1])
+		h += uint32(data[len(data)-1])
 		h *= m
 		h ^= (h >> r)
 	}
